deleting tasks endpoint working

+++ back-end/service/task.service.ts
@@ -1,6 +1,8 @@
+import { ta } from 'date-fns/locale';
 import { Task } from '../model/task';
 import taskDb from '../repository/task.db';
 import userDb from '../repository/user.db';
+import columnDb from '../repository/column.db';
 
 const getAllTasks = () => {
     return taskDb.getTasks();
@@ -15,6 +17,16 @@ const createTask = (task: Task) => {
 }
 
 const deleteTask = (taskId: string) => {
+    const task = taskDb.getTaskById(taskId);
+    if (!task) {
+        throw new Error('Task not found');
+    }
+    const columnId = `column${taskId.split('-').splice(1).join('-')}`;
+    const column = columnDb.getColumnById(columnId);
+    if (!column) {
+        throw new Error('Column not found');
+    }
+    column.removeTask(taskId);
     taskDb.removeTask(taskId);
 }
